refactor visualization directory
